Speed-up reference matching
The time needed to find and link references scale linearly with the
number of references and can be real slow for larger projects. So this
is a worthwhile focus point for performance improvements. This commit
alters the implementation to use just one big regex pattern that matches
all references, which appers to be about 100% faster for larger
projects. For smaller projects the gains are smaller, but still
positive.